request log for graphql
