+++ packages/tc-api-express/lib/request-log.js
@@ -1,0 +1,15 @@
+	logger,
+	setContext,
+} = require('@financial-times/tc-api-express-logger');
+const requestLog = (endpointName, method, req) => {
+	setContext({
+		endpointName,
+		method,
+		params: req.params,
+		query: req.query,
+		bodyKeys: Object.keys(req.body || {}),
+	});
+	logger.info(`[APP] ${endpointName} ${method}`);
+};
+
+module.exports = {requestLog}
